+++ packages/types/src/scope.ts
@@ -2,6 +2,7 @@ import { Breadcrumb } from './breadcrumb';
 import { EventProcessor } from './eventprocessor';
 import { Severity } from './severity';
 import { Span } from './span';
+import { Transaction } from './transaction';
 import { User } from './user';
 
 /** JSDocs */
@@ -71,10 +72,9 @@ export interface Scope {
   setLevel(level: Severity): this;
 
   /**
-   * Sets the transaction on the scope for future events.
-   * @param transaction string This will be converted in a tag in Sentry
+   * Sets the transaction name on the scope for future events.
    */
-  setTransaction(transaction?: string): this;
+  setTransactionName(name?: string): this;
 
   /**
    * Sets context data with the given name.
@@ -88,6 +88,16 @@ export interface Scope {
    * @param span Span
    */
   setSpan(span?: Span): this;
+
+  /**
+   * Returns the `Span` if there is one
+   */
+  getSpan(): Span | undefined;
+
+  /**
+   * Returns the `Transaction` if there is one
+   */
+  getTransaction(): Transaction | undefined;
 
   /**
    * Updates the scope with provided data. Can work in three variations: